Added add item to inventory changed model for inventory as well
Added add item to inventory Andre helped me make gui for the adding page and I changed the database for inventory and the model for it to show the stock limit and quantity for specific restaurants. I fixed the inventory options page to show the restaurants items and added a chef required wrap which admins can have access to.